+++ src/seppl/io/__init__.py
@@ -1,4 +1,4 @@
-from ._reader import Reader, DirectReader
+from ._reader import Reader, DirectReader, InfiniteReader
 from ._writer import Writer, DirectWriter, StreamWriter, DirectStreamWriter, BatchWriter, DirectBatchWriter, DataCollector
 from ._filter import BatchFilter, StreamFilter, FILTER_ACTIONS, FILTER_ACTION_KEEP, FILTER_ACTION_DISCARD, filter_data, FilterPipelineIterator, MultiFilter
 from ._filter import BatchFilter as Filter  # for backwards compatibility