+++ src/seppl/__init__.py
@@ -4,6 +4,7 @@ from ._plugin import AliasSupporter, get_aliases, has_aliases, get_all_names
 from ._plugin import SkippablePlugin, add_skip_option
 from ._args import split_cmdline, split_args, args_to_objects, is_help_requested, enumerate_plugins, escape_args, unescape_args, resolve_handler, load_args, save_args
 from ._registry import Registry, MODE_DYNAMIC, MODE_EXPLICIT, MODES
+from ._class_cache import ClassCache
 from ._class_registry import ClassListerRegistry, get_class_lister, DEFAULT
 from ._entry_points import generate_entry_points
 from ._help import generate_plugin_usage, generate_help, HELP_FORMATS, HELP_FORMAT_TEXT, HELP_FORMAT_MARKDOWN