+++ CHANGES.rst
@@ -20,6 +20,7 @@ Changelog
 - added `filter_data` generator function and `FilterPipelineIterator` iterator class for
   efficiently process data
 - the `classes_to_str` and `get_class_name` methods can remove `builtins.` via the clean=True parameter now
+- introduced the `InifiniteReader` mixin which automatically disables batch mode
 
 
 0.2.21 (2025-08-19)